Major refactoring and feature additions: Pipeline modularization, legacy code archival, and production readiness
## Core Architecture Improvements

### Pipeline Refactoring
- Migrated pipeline to modular structure (pipeline/ directory)
  - Separated stages into pipeline/stages.rs
  - Added pipeline/cache.rs for caching layer
  - Created pipeline/metrics.rs for metrics collection
  - Implemented pipeline/core.rs with improved state management
  - Added pipeline_legacy.rs shim for backward compatibility

### Code Organization & Cleanup
- Archived legacy code to archive/ directory:
  - Moved *.full backup files (config.rs.full, learning.rs.full, pipeline.rs.full)
  - Archived old pipeline_v2 implementation
  - Moved legacy test files to archive/legacy/tests/
  - Relocated legacy source files to archive/legacy/src/
  - Created archive/README.md documenting archival rationale
- Removed obsolete test files (test_cpp_config, test_extension, test_qwen*.py, etc.)
- Cleaned up redundant configuration files

## New Features & Components

### Reliability & Resilience
- Added circuit_breaker.rs for service failure handling
- Implemented degradation_tiers.rs for graceful degradation
- Created health.rs for health check endpoints
- Added resource_budget.rs for resource management
- Implemented security.rs for security enhancements

### Temporal Topological Data Analysis
- Added temporal_tda.rs for temporal topological analysis
- Created temporal_tda_test_suite/ with comprehensive test coverage
- Added federated_tda_tests.rs for distributed TDA scenarios
- Implemented Python wrapper (giotto_tda_wrapper.py) for TDA operations

### Testing Infrastructure
- Added new test binaries:
  - resource_test.rs for resource budget testing
  - soak_prompts_v2.rs and soak_test_v2.rs for improved soak testing
  - temporal_tda_test.rs for TDA validation
- Enhanced existing test suites with better coverage

## Deployment & Infrastructure

### Kubernetes Deployment
- Added Helm charts (deployment/helm/niodoo/)
  - ConfigMap templates for configuration management
  - Values.yaml with comprehensive defaults
- Created Kubernetes deployment manifests (deployment/k8s/)
- Added monitoring setup:
  - Grafana dashboard configuration
  - Prometheus alert rules

### Docker Support
- Added Dockerfile for containerized deployment
- Created .dockerignore for optimized builds

## Configuration & Dependencies

### Build System Updates
- Updated Cargo.toml files across modules
- Added ONNX Runtime GPU libraries (v1.18.1) to third_party/
- Updated service startup scripts (start_all_services.sh, check_all_services.sh)

### Configuration Management
- Added phase5_config.yaml configuration file
- Updated config.rs with improved configuration handling
- Added tcs_runtime.env for runtime environment variables

## Documentation

### Updated Documentation
- Enhanced README.md with latest architecture details
- Comprehensive CHANGELOG.md updates documenting all changes
- Added archive/README.md explaining code archival strategy

## Component Updates

### Core Components Modified
- Updated 50+ source files across niodoo_real_integrated/src/
- Improved API clients, embedding, ERAG, generation, curator modules
- Enhanced MCTS, learning, memory, and topology components
- Updated evaluation metrics and synthetic data generation

### Integration Improvements
- Improved Qwen integration components
- Enhanced RAG pipeline (retrieval, generation, local embeddings)
- Updated token promotion and dynamic tokenizer
- Improved consciousness engine and personal memory modules

## Testing & Quality Assurance

### Test Suite Updates
- Updated concurrent_cascade_smoke_test.rs
- Enhanced soak_test.rs with improved scenarios
- Maintained compatibility with existing test infrastructure

## Build & Runtime

- Verified compilation success across all modules
- Fixed compilation errors and warnings
- Maintained backward compatibility where possible
- Added workspace bashrc and cursor environment files

## Summary

This commit represents a major milestone in codebase organization and production readiness:
- ✅ Modular pipeline architecture
- ✅ Legacy code properly archived
- ✅ Production deployment infrastructure
- ✅ Enhanced reliability and resilience features
- ✅ Comprehensive testing infrastructure
- ✅ Improved documentation

All changes maintain backward compatibility and have been verified to compile successfully.

+++ src/consciousness_engine/mod.rs
@@ -51,12 +51,12 @@ use crate::personal_memory::{PersonalConsciousnessStats, PersonalInsight, Person
 use crate::personality::{PersonalityManager, PersonalityType};
 use crate::qt_mock::QtEmotionBridge;
 use crate::soul_resonance::SoulResonanceEngine;
-use crate::token_promotion::dynamic_tokenizer::Tokenizer;
 use crate::token_promotion::{
     ConsensusEngine, DynamicTokenizer, PatternDiscoveryEngine, PromotionConfig,
     TokenPromotionEngine,
 };
 use std::sync::atomic::{AtomicU64, Ordering};
+use tokenizers::Tokenizer;
 
 // Import from submodules
 use self::brain_coordination::BrainCoordinator;
@@ -1617,7 +1617,11 @@ impl PersonalNiodooConsciousness {
 
         // Use emotional_weight to determine emotion type
         let avg_weight =
-            memories.iter().map(|m| m.emotional_weight).sum::<f64>() / memories.len() as f64;
+            memories
+                .iter()
+                .map(|m| m.emotional_weight as f64)
+                .sum::<f64>()
+                / memories.len() as f64;
 
         // Map weight to emotion type
         if avg_weight > 0.7 {
